Login auth bugfix2

+++ tola/pipeline.py
@@ -28,17 +28,20 @@ def social_user_tola(backend, uid, user=None, *args, **kwargs):
     """extension of the social user lookup in the pipeline to clear bad associations"""
     # try to get user from social auth storage:
     provider = backend.name
-    social = backend.strategy.storage.user.get_social_auth(provider, uid)
-    if social and social.user.email.lower() != uid.lower():
-        # we found a match, but the emails are different, delete the bad data:
-        social.delete()
+    if provider == 'google-oauth2':
+        social = backend.strategy.storage.user.get_social_auth(provider, uid)
+        if social and social.user.email.lower() != uid.lower():
+            # we found a match, but the emails are different, delete the bad data:
+            social.delete()
+            user = None
     # call the original social_user now that we know bad data has been expunged:
-    return social_user(backend, uid, user, *args, **kwargs)
+    social = social_user(backend, uid, user, *args, **kwargs)
+    return social
 
 
 def associate_user_tola(backend, uid, user=None, social=None, *args, **kwargs):
     """extension of the user-association step of the pipeline to avoid associating to previously logged-in users"""
-    if user and user.email and user.email.lower() != uid.lower():
+    if backend.name == 'google-oauth2' and user and user.email and user.email.lower() != uid.lower():
         # if we have a user logged in (sessioning!) and they don't match the uid entered
         # zero out the user so we don't associate badly a user with a bad social auth:
         user = None
